MINOR: extracted the majority of the logic from MemoryRecordsBuilder to an abstract class.
